+++ bittytax/conv/parsers/hitbtc.py
@@ -9,6 +9,35 @@ from ..exceptions import UnexpectedTypeError
 
 WALLET = "HitBTC"
 
+def parse_hitbtc_trades2(data_row, parser, _filename):
+    in_row = data_row.in_row
+    data_row.timestamp = DataParser.parse_timestamp(in_row[1])
+
+    if in_row[5] == "buy":
+        data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_TRADE,
+                                                 data_row.timestamp,
+                                                 buy_quantity=in_row[6],
+                                                 buy_asset=in_row[2].split('/')[1],
+                                                 sell_quantity=Decimal(in_row[8]) - \
+                                                               Decimal(in_row[10]),
+                                                 sell_asset=in_row[2].split('/')[2],
+                                                 fee_quantity=in_row[9],
+                                                 fee_asset=in_row[2].split('/')[2],
+                                                 wallet=WALLET)
+    elif in_row[5] == "sell":
+        data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_TRADE,
+                                                 data_row.timestamp,
+                                                 buy_quantity=Decimal(in_row[8]) + \
+                                                              Decimal(in_row[10]),
+                                                 buy_asset=in_row[2].split('/')[2],
+                                                 sell_quantity=in_row[6],
+                                                 sell_asset=in_row[2].split('/')[1],
+                                                 fee_quantity=in_row[9],
+                                                 fee_asset=in_row[2].split('/')[2],
+                                                 wallet=WALLET)
+    else:
+        raise UnexpectedTypeError(5, parser.in_header[5], in_row[5])
+
 def parse_hitbtc_trades(data_row, parser, _filename):
     in_row = data_row.in_row
     data_row.timestamp = DataParser.parse_timestamp(in_row[0])
@@ -38,6 +67,24 @@ def parse_hitbtc_trades(data_row, parser, _filename):
     else:
         raise UnexpectedTypeError(4, parser.in_header[4], in_row[4])
 
+
+def parse_hitbtc_deposits_withdrawals2(data_row, _parser, _filename):
+    in_row = data_row.in_row
+    data_row.timestamp = DataParser.parse_timestamp(in_row[1])
+
+    if in_row[3] == "Withdraw":
+        data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_WITHDRAWAL,
+                                                 data_row.timestamp,
+                                                 sell_quantity=abs(Decimal(in_row[4])),
+                                                 sell_asset=in_row[7],
+                                                 wallet=WALLET)
+    elif in_row[3] == "Deposit":
+        data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_DEPOSIT,
+                                                 data_row.timestamp,
+                                                 buy_quantity=in_row[4],
+                                                 buy_asset=in_row[7],
+                                                 wallet=WALLET)
+
 def parse_hitbtc_deposits_withdrawals(data_row, _parser, _filename):
     in_row = data_row.in_row
     data_row.timestamp = DataParser.parse_timestamp(in_row[0])
@@ -57,10 +104,24 @@ def parse_hitbtc_deposits_withdrawals(data_row, _parser, _filename):
 
 DataParser(DataParser.TYPE_EXCHANGE,
            "HitBTC",
+           ['Email', 'Date (UTC)', 'Instrument', 'Trade ID', 'Order ID', 'Side', 'Quantity',
+            'Price', 'Volume', 'Fee', 'Rebate', 'Total'],
+           worksheet_name="HitBTC T",
+           row_handler=parse_hitbtc_trades2)
+
+DataParser(DataParser.TYPE_EXCHANGE,
+           "HitBTC",
            ['Date (UTC)', 'Instrument', 'Trade ID', 'Order ID', 'Side', 'Quantity', 'Price',
             'Volume', 'Fee', 'Rebate', 'Total'],
            worksheet_name="HitBTC T",
            row_handler=parse_hitbtc_trades)
+
+DataParser(DataParser.TYPE_EXCHANGE,
+           "HitBTC",
+           ['Email', 'Date (UTC)', 'Operation id', 'Type', 'Amount', 'Transaction hash',
+           'Main account balance', 'Currency'],
+           worksheet_name="HitBTC D,W",
+           row_handler=parse_hitbtc_deposits_withdrawals2)
 
 DataParser(DataParser.TYPE_EXCHANGE,
            "HitBTC",